+++ stsynphot/catalog.py
@@ -3,6 +3,7 @@
 
 # STDLIB
 import numbers
+from urllib.error import HTTPError
 
 # THIRD-PARTY
 import numpy as np
@@ -123,7 +124,9 @@ def get_catalog_index(gridname):
     Parameters
     ----------
     gridname : str
-        See :func:`grid_to_spec`.
+        See :func:`grid_to_spec` for supported grids.
+        Any other input is assumed to be custom grid stored
+        in ``crgrid$`` as expected by :func:`~stsynphot.stio.irafconvert`.
 
     Returns
     -------
@@ -141,18 +144,21 @@ def get_catalog_index(gridname):
     elif gridname == 'phoenix':
         catdir = 'crgridphoenix$'
     else:
-        raise synexceptions.SynphotError(
-            f'{gridname} is not a supported catalog grid.')
+        catdir = 'crgrid${}'.format(gridname)
 
     catdir = stio.irafconvert(catdir)
     filename = stio.resolve_filename(catdir, 'catalog.fits')
 
     # If not cached, read from grid catalog and cache it
     if filename not in _CACHE:
-        data = stio.read_catalog(filename)  # EXT 1
-        _CACHE[filename] = [list(map(float, index.split(','))) +
-                            [data['FILENAME'][i]]
-                            for i, index in enumerate(data['INDEX'])]
+        try:
+            data = stio.read_catalog(filename)  # EXT 1
+            _CACHE[filename] = [list(map(float, index.split(','))) +
+                                [data['FILENAME'][i]]
+                                for i, index in enumerate(data['INDEX'])]
+        except (FileNotFoundError, HTTPError) as e:
+            raise synexceptions.SynphotError(
+                f'{gridname} catalog was not found: {e}') from e
 
     return _CACHE[filename], catdir